Move material buttons parent component to the reserve physical folder
